+++ tests/test_igzip_lib.py
@@ -1,22 +1,13 @@
-# Copyright (c) 2020 Leiden University Medical Center
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in
-# all copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
+# Copyright (c) 2001, 2002, 2003, 2004, 2005, 2006, 2007, 2008, 2009, 2010,
+# 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2020, 2021, 2022
+# Python Software Foundation; All Rights Reserved
+
+# This file is part of python-isal which is distributed under the
+# PYTHON SOFTWARE FOUNDATION LICENSE VERSION 2.
+
+# TestIgzipDecompressor was modified from TestBZ2Decompress from Cpython's
+# Lib/test/test_bz2.py.
+
 import gzip
 import itertools
 import os
@@ -31,6 +22,7 @@ from isal.igzip_lib import (
     DECOMP_ZLIB, DECOMP_ZLIB_NO_HDR, DECOMP_ZLIB_NO_HDR_VER, MEM_LEVEL_DEFAULT,
     MEM_LEVEL_EXTRA_LARGE, MEM_LEVEL_LARGE, MEM_LEVEL_MEDIUM, MEM_LEVEL_MIN,
     MEM_LEVEL_SMALL)
+from isal.igzip_lib import IgzipDecompressor
 
 import pytest
 
@@ -79,12 +71,12 @@ class TestIgzipDecompressor():
     BAD_DATA = b"Not a valid deflate block"
 
     def test_decompress(self):
-        decomp = igzip_lib.decompressobj()
+        decomp = IgzipDecompressor()
         decompressed = decomp.decompress(self.DATA)
         assert decompressed == self.TEXT
 
     def testDecompressChunks10(self):
-        igzd = igzip_lib.decompressobj()
+        igzd = IgzipDecompressor()
         text = b''
         n = 0
         while True:
@@ -96,14 +88,14 @@ class TestIgzipDecompressor():
         assert text == self.TEXT
 
     def testDecompressUnusedData(self):
-        igzd = igzip_lib.decompressobj()
+        igzd = IgzipDecompressor()
         unused_data = b"this is unused data"
         text = igzd.decompress(self.DATA+unused_data)
         assert text == self.TEXT
         assert igzd.unused_data == unused_data
 
     def testEOFError(self):
-        igzd = igzip_lib.decompressobj()
+        igzd = IgzipDecompressor()
         igzd.decompress(self.DATA)
         with pytest.raises(EOFError):
             igzd.decompress(b"anything")
@@ -119,7 +111,7 @@ class TestIgzipDecompressor():
         try:
             data = block * (size // blocksize + 1)
             compressed = igzip_lib.compress(data)
-            igzd = igzip_lib.decompressobj()
+            igzd = IgzipDecompressor()
             decompressed = igzd.decompress(compressed)
             assert decompressed == data
         finally:
@@ -130,10 +122,10 @@ class TestIgzipDecompressor():
     def testPickle(self):
         for proto in range(pickle.HIGHEST_PROTOCOL + 1):
             with pytest.raises(TypeError):
-                pickle.dumps(igzip_lib.decompressobj(), proto)
+                pickle.dumps(IgzipDecompressor(), proto)
 
     def testDecompressorChunksMaxsize(self):
-        igzd = igzip_lib.decompressobj()
+        igzd = IgzipDecompressor()
         max_length = 100
         out = []
 
@@ -166,7 +158,7 @@ class TestIgzipDecompressor():
     def test_decompressor_inputbuf_1(self):
         # Test reusing input buffer after moving existing
         # contents to beginning
-        igzd = igzip_lib.decompressobj()
+        igzd = IgzipDecompressor()
         out = []
 
         # Create input buffer and fill it
@@ -187,7 +179,7 @@ class TestIgzipDecompressor():
     def test_decompressor_inputbuf_2(self):
         # Test reusing input buffer by appending data at the
         # end right away
-        igzd = igzip_lib.decompressobj()
+        igzd = IgzipDecompressor()
         out = []
 
         # Create input buffer and empty it
@@ -207,7 +199,7 @@ class TestIgzipDecompressor():
     def test_decompressor_inputbuf_3(self):
         # Test reusing input buffer after extending it
 
-        igzd = igzip_lib.decompressobj()
+        igzd = IgzipDecompressor()
         out = []
 
         # Create almost full input buffer
@@ -221,7 +213,7 @@ class TestIgzipDecompressor():
         assert b''.join(out) == self.TEXT
 
     def test_failure(self):
-        igzd = igzip_lib.decompressobj()
+        igzd = IgzipDecompressor()
         with pytest.raises(Exception):
             igzd.decompress(self.BAD_DATA * 30)
         # Make sure there are no internal consistencies
@@ -236,7 +228,7 @@ def test_igzip_decompressor_raw_deflate_unused_data_zlib(test_offset):
     trailer = data[-4:]
     raw_deflate_incomplete_trailer = no_header[:-test_offset]
     true_unused_data = trailer[:-test_offset]
-    igzd = igzip_lib.decompressobj(flag=DECOMP_DEFLATE)
+    igzd = IgzipDecompressor(flag=DECOMP_DEFLATE)
     igzd.decompress(raw_deflate_incomplete_trailer)
     if igzd.eof:
         assert igzd.unused_data == true_unused_data
@@ -249,7 +241,7 @@ def test_igzip_decompressor_raw_deflate_unused_data_gzip(test_offset):
     trailer = data[-8:]
     raw_deflate_incomplete_trailer = no_header[:-test_offset]
     true_unused_data = trailer[:-test_offset]
-    igzd = igzip_lib.decompressobj(flag=DECOMP_DEFLATE)
+    igzd = IgzipDecompressor(flag=DECOMP_DEFLATE)
     igzd.decompress(raw_deflate_incomplete_trailer)
     if igzd.eof:
         assert igzd.unused_data == true_unused_data
@@ -263,7 +255,7 @@ def test_decompression_flags(unused_size, flag_pair, data_size):
     unused_data = b"abcdefghijklmnopqrstuvwxyz"[:unused_size]
     data = RAW_DATA[:data_size]
     compressed = igzip_lib.compress(data, flag=comp_flag)
-    decompressor = igzip_lib.decompressobj(flag=decomp_flag)
+    decompressor = igzip_lib.IgzipDecompressor(flag=decomp_flag)
     result = decompressor.decompress(compressed + unused_data)
     assert result == data